Separate fetch plugin from path resolution.
