+++ src/plugins/unpkg-path-plugin.ts
@@ -1,10 +1,4 @@
 import * as esbuild from "esbuild-wasm";
-import axios from "axios";
-import localForage from "localforage";
-
-const fileCache = localForage.createInstance({
-  name: "filecache",
-});
 
 /**
  * Specifies actions to take when parsing url package information for requests
@@ -13,7 +7,7 @@ const fileCache = localForage.createInstance({
  * @returns An esbuild-plugin to add steps to onResolve and onLoad when the build
  *          function is called.
  */
-export const unpkgPathPlugin = (inputCode: string) => {
+export const unpkgPathPlugin = () => {
   return {
     name: "unpkg-path-plugin",
     setup(build: esbuild.PluginBuild) {
@@ -36,38 +30,6 @@ export const unpkgPathPlugin = (inputCode: string) => {
           namespace: "a",
           path: `https://unpkg.com/${args.path}`,
         };
-      });
-
-      build.onLoad({ filter: /.*/ }, async (args: any) => {
-        console.log("onLoad", args);
-
-        if (args.path === "index.js") {
-          return {
-            loader: "jsx",
-            contents: inputCode,
-          };
-        }
-
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
-          args.path
-        );
-
-        if (cachedResult) {
-          return cachedResult;
-        }
-
-        const { data, request } = await axios.get(args.path);
-
-        const result: esbuild.OnLoadResult = {
-          loader: "jsx",
-          contents: data,
-          resolveDir: new URL("./", request.responseURL).pathname,
-        };
-
-        // store response in cache using args.path as key
-        await fileCache.setItem(args.path, result);
-
-        return result;
       });
     },
   };
